Use executable instead of editing bashrc

+++ create.sh
@@ -20,15 +20,32 @@ fi
 
 $CONTAINER_TOOL build -t $1 --build-arg "NVIM_CONFIG_URL=$2" .
 $CONTAINER_TOOL volume rm --force $1
-$CONTAINER_TOOL volume create $1
+
+if [ -e "/usr/local/bin/$1" ]; then
+	# File exists, prompt the user if they want to overwrite it
+	read -p "File '/usr/local/bin/$1' already exists. Do you want to overwrite it? (y/n): " answer
+	if [ "$answer" != "y" ]; then
+		echo "Aborted. File not overwritten."
+		exit 0
+	fi
+fi
 
 # Add nvim alias to bashrc
-cat <<EOF | tee -a ~/.bashrc
+cat <<EOF | sudo tee /usr/local/bin/$1 >/dev/null
+#!/bin/bash
 
-function $1 () {
-  rm -f /tmp/container_id_$1
-  $CONTAINER_TOOL run --cidfile /tmp/container_id_$1 -v $1:/root/.local -v .:/root/dev -it $1 "\$@"
-  echo "Committing changes to docker container:"
-  $CONTAINER_TOOL container commit \$(cat /tmp/container_id_$1) $1
-}
+rm -f /tmp/container_id_$1
+$CONTAINER_TOOL volume create $1
+$CONTAINER_TOOL run \
+  --cidfile /tmp/container_id_$1 \
+  -v $1:/root/.local \
+  -v .:/root/dev \
+  -v ~/.gitconfig:/root/.gitconfig \
+  -it \
+  $1 \
+  "\$@"
+echo "Committing changes to docker container:"
+$CONTAINER_TOOL container commit \$(cat /tmp/container_id_$1) $1
 EOF
+
+sudo chmod +x /usr/local/bin/$1
